Move util functions from gatsby-node into separate file
