+++ gatsby-node.js
@@ -6,119 +6,21 @@ gracefulFs.gracefulify(realFs);
 const { createFilePath } = require(`gatsby-source-filesystem`);
 const { exec, execSync } = require('child_process');
 
+const {
+  sortVersionArray,
+  replacePathVersion,
+  filePathToDocType,
+  removeTrailingSlash,
+  isPathAnIndexPage,
+  removeNullEntries,
+  pathToDepth,
+  mdxNodesToTree,
+  buildProductVersions,
+  reportMissingIndex,
+} = require('./src/constants/gatsby-node-utils.js');
+
 const isBuild = process.env.NODE_ENV === 'production';
 const isProduction = process.env.APP_ENV === 'production';
-
-const sortVersionArray = (versions) => {
-  return versions
-    .map((version) => version.replace(/\d+/g, (n) => +n + 100000))
-    .sort()
-    .map((version) => version.replace(/\d+/g, (n) => +n - 100000));
-};
-
-const replacePathVersion = (path, version = 'latest') => {
-  const splitPath = path.split('/');
-  const postVersionPath = splitPath.slice(3).join('/');
-  return `/${splitPath[1]}/${version}${
-    postVersionPath.length > 0 ? `/${postVersionPath}` : '/'
-  }`;
-};
-
-const filePathToDocType = (filePath) => {
-  if (filePath.includes('/product_docs/')) {
-    return 'doc';
-  } else if (filePath.includes('/advocacy_docs/')) {
-    return 'advocacy';
-  } else {
-    return 'gh_doc';
-  }
-};
-
-const removeTrailingSlash = (url) => {
-  if (url.endsWith('/')) {
-    return url.slice(0, -1);
-  }
-  return url;
-};
-
-const isPathAnIndexPage = (filePath) =>
-  filePath.endsWith('/index.mdx') || filePath === 'index.mdx';
-
-const removeNullEntries = (obj) => {
-  if (!obj) return obj;
-  for (const [key, value] of Object.entries(obj)) {
-    if (value == null) delete obj[key];
-  }
-  return obj;
-};
-
-const pathToDepth = (path) => {
-  return path.split('/').filter((s) => s.length > 0).length;
-};
-
-const mdxNodesToTree = (nodes) => {
-  const buildNode = (path, parent) => {
-    return {
-      path: path,
-      parent: parent,
-      children: [],
-      mdxNode: null,
-      depth: pathToDepth(path),
-    };
-  };
-
-  const rootNode = buildNode('/', null);
-
-  const findOrInsertNode = (currentNode, path) => {
-    const node = currentNode.children.find((child) => child.path === path);
-    if (node) return node;
-
-    const newNode = buildNode(path, currentNode);
-    currentNode.children.push(newNode);
-    return newNode;
-  };
-
-  const addNode = (node) => {
-    const splitPath = node.fields.path.split('/');
-    let currentNode = rootNode;
-    for (let i = 2; i < splitPath.length; i++) {
-      const path = `/${splitPath.slice(1, i).join('/')}/`;
-      currentNode = findOrInsertNode(currentNode, path);
-      if (path === node.fields.path) {
-        currentNode.mdxNode = node;
-      }
-    }
-  };
-
-  nodes.forEach((node) => addNode(node));
-
-  return rootNode;
-};
-
-const buildProductVersions = (nodes) => {
-  const versionIndex = {};
-
-  nodes.forEach((node) => {
-    const { docType, product, version } = node.fields;
-    if (docType !== 'doc') return;
-
-    if (!versionIndex[product]) {
-      versionIndex[product] = [version];
-    } else {
-      if (!versionIndex[product].includes(version)) {
-        versionIndex[product].push(version);
-      }
-    }
-  });
-
-  for (const product in versionIndex) {
-    versionIndex[product] = sortVersionArray(versionIndex[product]).reverse();
-  }
-
-  return versionIndex;
-};
-
-//////////////////////////////////////////////////////////////
 
 exports.onCreateNode = async ({ node, getNode, actions }) => {
   const { createNodeField } = actions;
@@ -238,10 +140,10 @@ exports.createPages = async ({ actions, graphql, reporter }) => {
 
     // exit here if we're not dealing with an actual page
     if (!curr.mdxNode) {
-      // console.log(curr.path);
-
+      reportMissingIndex(reporter, curr);
       continue;
     }
+
     const node = curr.mdxNode;
 
     // set computed frontmatter
@@ -250,11 +152,6 @@ exports.createPages = async ({ actions, graphql, reporter }) => {
       ...frontmatterStack,
       ...[removeNullEntries(node.frontmatter)],
     );
-
-    // warn if no title
-    // if (!node.frontmatter.title) {
-    //   reporter.warn(`${node.fields.path} has no title`);
-    // }
 
     // set up frontmatter redirects
     if (node.frontmatter.redirects) {
@@ -277,29 +174,6 @@ exports.createPages = async ({ actions, graphql, reporter }) => {
       createGHDoc(node, gh_docs, actions);
     }
   }
-
-  // const folderIndex = {};
-
-  // nodes.forEach((doc) => {
-  //   const { path } = doc.fields;
-
-  //   const splitPath = path.split('/');
-  //   const subPath = splitPath.slice(0, splitPath.length - 2).join('/') + '/';
-  //   const { fileAbsolutePath } = doc;
-  //   if (fileAbsolutePath.includes('index.mdx')) {
-  //     folderIndex[path] = true;
-  //   } else {
-  //     if (!folderIndex[subPath]) {
-  //       folderIndex[subPath] = false;
-  //     }
-  //   }
-  // });
-
-  // for (let key of Object.keys(folderIndex)) {
-  //   if (!folderIndex[key]) {
-  //     reporter.warn(key + ' has no index.mdx');
-  //   }
-  // }
 };
 
 const createDoc = (doc, productVersions, docs, actions) => {
